added import BrowserRouter, Switch and Route. Trying to figure out slug
